@@ -1,3 +1,3 @@
 module github.com/mertenvg/logr
 
-go 1.13
+go 1.18
