@@ -1,7 +1,7 @@
 package logr
 
 func init() {
-	pool.New = func() interface{} {
+	pool.New = func() any {
 		return &Message{}
 	}
 	SetBufferSize(10000)
