@@ -17,7 +17,7 @@ var (
 )
 
 // SetMeta sets the global meta data attached to every log message
-func SetMeta(data map[string]interface{}) {
+func SetMeta(data map[string]any) {
 	logr = logr.With(data)
 }
 
@@ -44,96 +44,93 @@ func Wait() {
 
 // Logger defines the methods available both by the logr package and Logr containing additional meta data.
 type Logger interface {
-	Panic(v ...interface{})
-	Panicf(msg string, v ...interface{})
-	Error(v ...interface{}) string
-	Errorf(msg string, v ...interface{}) string
-	Warn(v ...interface{}) string
-	Warnf(msg string, v ...interface{}) string
-	Info(v ...interface{}) string
-	Infof(msg string, v ...interface{}) string
-	Debug(v ...interface{}) string
-	Debugf(msg string, v ...interface{}) string
-	Success(v ...interface{}) string
-	Successf(msg string, v ...interface{}) string
-	With(data map[string]interface{}) Logger
+	Panic(v ...any)
+	Panicf(msg string, v ...any)
+	Error(v ...any) string
+	Errorf(msg string, v ...any) string
+	Warn(v ...any) string
+	Warnf(msg string, v ...any) string
+	Info(v ...any) string
+	Infof(msg string, v ...any) string
+	Debug(v ...any) string
+	Debugf(msg string, v ...any) string
+	Success(v ...any) string
+	Successf(msg string, v ...any) string
+	With(data Meta) Logger
 }
 
 // Logr implements the Logger interface
 type Logr struct {
-	meta map[string]interface{}
+	meta Meta
 }
 
 // Panic logs inputs as panics and panics
-func (l *Logr) Panic(v ...interface{}) {
+func (l *Logr) Panic(v ...any) {
 	code := log(P, true, Interfaces(v).SSV(), l.meta)
 	panic(code)
 }
 
 // Panicf logs a formatted message as a panic and panics
-func (l *Logr) Panicf(msg string, v ...interface{}) {
+func (l *Logr) Panicf(msg string, v ...any) {
 	code := logf(P, true, msg, v, l.meta)
 	panic(code)
 }
 
 // Error logs inputs as errors
-func (l *Logr) Error(v ...interface{}) string {
+func (l *Logr) Error(v ...any) string {
 	return log(E, false, Interfaces(v).SSV(), l.meta)
 }
 
 // Errorf logs a formatted message as an error
-func (l *Logr) Errorf(msg string, v ...interface{}) string {
+func (l *Logr) Errorf(msg string, v ...any) string {
 	return logf(E, false, msg, v, l.meta)
 }
 
 // Warn logs inputs as warnings
-func (l *Logr) Warn(v ...interface{}) string {
+func (l *Logr) Warn(v ...any) string {
 	return log(W, false, Interfaces(v).SSV(), l.meta)
 }
 
 // Warnf logs a formatted message as a warning
-func (l *Logr) Warnf(msg string, v ...interface{}) string {
+func (l *Logr) Warnf(msg string, v ...any) string {
 	return logf(W, false, msg, v, l.meta)
 }
 
 // Info logs inputs as info messages
-func (l *Logr) Info(v ...interface{}) string {
+func (l *Logr) Info(v ...any) string {
 	return log(I, false, Interfaces(v).SSV(), l.meta)
 }
 
 // Infof logs a formatted message as an info message
-func (l *Logr) Infof(msg string, v ...interface{}) string {
+func (l *Logr) Infof(msg string, v ...any) string {
 	return logf(I, false, msg, v, l.meta)
 }
 
 // Debug logs inputs as debug messages
-func (l *Logr) Debug(v ...interface{}) string {
+func (l *Logr) Debug(v ...any) string {
 	return log(D, false, Interfaces(v).SSV(), l.meta)
 }
 
 // Debugf logs a formatted message as a debug message
-func (l *Logr) Debugf(msg string, v ...interface{}) string {
+func (l *Logr) Debugf(msg string, v ...any) string {
 	return logf(D, false, msg, v, l.meta)
 }
 
 // Success logs inputs as success messages
-func (l *Logr) Success(v ...interface{}) string {
+func (l *Logr) Success(v ...any) string {
 	return log(S, false, Interfaces(v).SSV(), l.meta)
 }
 
 // Successf logs a formatted message as a success message
-func (l *Logr) Successf(msg string, v ...interface{}) string {
+func (l *Logr) Successf(msg string, v ...any) string {
 	return logf(S, false, msg, v, l.meta)
 }
 
 // With metadata in the log messages
-func (l *Logr) With(data map[string]interface{}) Logger {
-	meta := make(map[string]interface{})
-	for k, v := range l.meta {
-		meta[k] = v
-	}
+func (l *Logr) With(data Meta) Logger {
+	meta := l.meta.Copy()
 	for k, v := range data {
-		meta[k] = v
+		meta.With(k, v)
 	}
 	return &Logr{
 		meta: meta,
@@ -141,12 +138,12 @@ func (l *Logr) With(data map[string]interface{}) Logger {
 }
 
 // format a msg and log as given type
-func logf(t Type, wait bool, msg string, args []interface{}, meta map[string]interface{}) string {
+func logf(t Type, wait bool, msg string, args []any, meta map[string]any) string {
 	return log(t, wait, fmt.Sprintf(msg, args...), meta)
 }
 
 // log inputs to given type
-func log(t Type, wait bool, msg string, meta map[string]interface{}) string {
+func log(t Type, wait bool, msg string, meta map[string]any) string {
 	now := time.Now()
 	m := pool.Get().(*Message)
 	m.Type = t
@@ -165,66 +162,66 @@ func log(t Type, wait bool, msg string, meta map[string]interface{}) string {
 }
 
 // Panic logs inputs as panics and panics
-func Panic(v ...interface{}) {
+func Panic(v ...any) {
 	logr.Panic(v...)
 }
 
 // Panicf logs a formatted message as a panic and panics
-func Panicf(msg string, v ...interface{}) {
+func Panicf(msg string, v ...any) {
 	logr.Panicf(msg, v...)
 }
 
 // Error logs inputs as errors
-func Error(v ...interface{}) string {
+func Error(v ...any) string {
 	return logr.Error(v...)
 }
 
 // Errorf logs a formatted message as an error
-func Errorf(msg string, v ...interface{}) string {
+func Errorf(msg string, v ...any) string {
 	return logr.Errorf(msg, v...)
 }
 
 // Warn logs inputs as warnings
-func Warn(v ...interface{}) string {
+func Warn(v ...any) string {
 	return logr.Warn(v...)
 }
 
 // Warnf logs a formatted message as a warning
-func Warnf(msg string, v ...interface{}) string {
+func Warnf(msg string, v ...any) string {
 	return logr.Warnf(msg, v...)
 }
 
 // Info logs inputs as info messages
-func Info(v ...interface{}) string {
+func Info(v ...any) string {
 	return logr.Info(v...)
 }
 
 // Infof logs a formatted message as an info message
-func Infof(msg string, v ...interface{}) string {
+func Infof(msg string, v ...any) string {
 	return logr.Infof(msg, v...)
 }
 
 // Debug logs inputs as debug messages
-func Debug(v ...interface{}) string {
+func Debug(v ...any) string {
 	return logr.Debug(v...)
 }
 
 // Debugf logs a formatted message as a debug message
-func Debugf(msg string, v ...interface{}) string {
+func Debugf(msg string, v ...any) string {
 	return logr.Debugf(msg, v...)
 }
 
 // Success logs inputs as success messages
-func Success(v ...interface{}) string {
+func Success(v ...any) string {
 	return logr.Success(v...)
 }
 
 // Successf logs a formatted message as a success message
-func Successf(msg string, v ...interface{}) string {
+func Successf(msg string, v ...any) string {
 	return logr.Successf(msg, v...)
 }
 
 // With metadata in the log messages
-func With(data map[string]interface{}) Logger {
+func With(data Meta) Logger {
 	return logr.With(data)
 }
